split the models defined in app.py into smaller files
