+++ models/error.py
@@ -1,0 +1,10 @@
+from . import db
+
+class Error(db.Model):
+    '''
+        Errors from run_update.py
+    '''
+    # Columns
+    id = db.Column(db.Integer(), primary_key=True)
+    error = db.Column(db.Unicode())
+    time = db.Column(db.DateTime(False))
