+++ name_utils.py
@@ -1,0 +1,20 @@
+
+
+def is_safe_name(name):
+    ''' Return True if the string is a safe name.
+    '''
+    return raw_name(safe_name(name)) == name
+
+def safe_name(name):
+    ''' Return URL-safe organization name with spaces replaced by dashes.
+
+        Slashes will be removed, which is incompatible with raw_name().
+    '''
+    return name.replace(' ', '-').replace('/', '-').replace('?', '-').replace('#', '-')
+
+def raw_name(name):
+    ''' Return raw organization name with dashes replaced by spaces.
+
+        Also replace old-style underscores with spaces.
+    '''
+    return name.replace('_', ' ').replace('-', ' ')
