+++ run_update.py
@@ -17,7 +17,8 @@ from dateutil.tz import tzoffset
 import feedparser
 
 from feeds import get_first_working_feed_link
-from app import db, Project, Organization, Story, Event, Error, Issue, Label, is_safe_name
+from models import db, Project, Organization, Story, Event, Error, Issue, Label
+from name_utils import is_safe_name
 
 # Logging Setup
 logging.basicConfig(level=logging.INFO)
